# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shan Loray Beauty E-Commerce API",
        "description": "Complete REST API for the Shan Loray luxury beauty e-commerce platform",
        "contact": {
            "name": "Shan Loray API Support",
            "email": "api@shanloray.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://shanloray.com/terms"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 Server"
        }
    ],
    "paths": {
...
        "/consultations": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "List user's consultations",
                "operationId": "03192b1632d51781c7469b20e2771020",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "upcoming",
                        "in": "query",
                        "description": "Upcoming only",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of consultations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "consultations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Consultation"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Book a new consultation",
                "operationId": "8814cd9ef86cdeb5714a8f93ce171ece",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "specialist_id",
                                    "scheduled_date",
                                    "scheduled_time"
                                ],
                                "properties": {
                                    "specialist_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "video",
                                            "in_person",
                                            "chat"
                                        ],
                                        "example": "video"
                                    },
                                    "scheduled_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-02-10"
                                    },
                                    "scheduled_time": {
                                        "type": "string",
                                        "example": "10:00"
                                    },
                                    "concern_area": {
                                        "type": "string"
                                    },
                                    "customer_notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consultation booked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Consultation booked successfully"
                                        },
                                        "consultation": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Time slot not available"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/consultations/{consultation}": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Get a specific consultation",
                "operationId": "1d0250c56d66d078a0d963a58fa0394d",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consultation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "consultation": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consultation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/consultations/{consultation}/cancel": {
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Cancel a consultation",
                "operationId": "2f037e74ba93c82f8af34c31f8549fbc",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consultation cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "consultation": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cannot be cancelled"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/consultations/{consultation}/reschedule": {
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Reschedule a consultation",
                "operationId": "0fd64c7773f19a664d4673d1f72bb3c4",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "scheduled_date",
                                    "scheduled_time"
                                ],
                                "properties": {
                                    "scheduled_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "scheduled_time": {
                                        "type": "string",
                                        "example": "14:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consultation rescheduled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "consultation": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cannot be rescheduled or time not available"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/consultations/{consultation}/rate": {
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Rate a completed consultation",
                "operationId": "f49082aab0d3d6bd90417222d3531386",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "review": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rating submitted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thank you for your feedback"
                                        },
                                        "consultation": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cannot be rated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Consultation": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "booking_reference": {
                        "type": "string",
                        "example": "CON2602061234ABC"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "scheduled_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "scheduled_time": {
                        "type": "string",
                        "example": "10:00"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 12
                    },
                    "total": {
                        "type": "integer",
                        "example": 56
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}